--- .github/workflows/install_git_hooks_windows.yml
@@ -1,44 +1,0 @@
-# THIS ACTION IS FOR VALIDATION WITHIN TEMPLATE REPO, AND WILL BE
-# REMOVED UPON RUNNING RENAMETEMPLATE GRADLE TASK.
-
-name: Install Windows Hooks
-
-# This will cancel any in progress workflows for the same PR, if
-# multiple pushes happen in quick succession.
-concurrency:
-  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
-  cancel-in-progress: true
-
-on: pull_request
-
-jobs:
-  install-windows-hooks:
-    runs-on: windows-latest
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Set Up JDK
-        uses: actions/setup-java@v4
-        with:
-          distribution: 'zulu'
-          java-version-file: .github/.java-version
-
-      - name: Setup Gradle
-        uses: gradle/gradle-build-action@v3
-
-      - name: Install
-        run: ./gradlew installGitHooks
-
-      - name: Check Pre-Commit File
-        uses: GuillaumeFalourd/assert-command-line-output@v2.4
-        with:
-          command_line: cat .git/hooks/pre-commit
-          assert_file_path: git-hooks/pre-commit-windows.sh
-          expected_result: PASSED
-
-      - name: Check Pre-Push File
-        uses: GuillaumeFalourd/assert-command-line-output@v2.4
-        with:
-          command_line: cat .git/hooks/pre-push
-          assert_file_path: git-hooks/pre-push-windows.sh
-          expected_result: PASSED
