+++ app/src/main/java/com/adammcneilly/spatialdialog/App.kt
@@ -1,0 +1,5 @@
+package com.adammcneilly.spatialdialog
+
+import android.app.Application
+
+class App : Application()
