+++ build.gradle.kts
@@ -3,10 +3,8 @@
 plugins {
     alias(libs.plugins.android.application).apply(false)
     alias(libs.plugins.benmanes.versions).apply(false)
-    alias(libs.plugins.cash.paparazzi).apply(false)
     alias(libs.plugins.compose.compiler).apply(false)
     alias(libs.plugins.detekt).apply(true) // Needs to be applied at the root, unlike others.
-    alias(libs.plugins.google.dagger.hilt).apply(false)
     alias(libs.plugins.google.ksp).apply(false)
     alias(libs.plugins.kotlin.android).apply(false)
     alias(libs.plugins.kotlinter).apply(false)
@@ -14,7 +12,6 @@ plugins {
 }
 
 apply(from = "buildscripts/githooks.gradle")
-apply(from = "buildscripts/setup.gradle")
 apply(from = "buildscripts/versionsplugin.gradle")
 
 subprojects {
