--- renovate.json
@@ -1,30 +1,0 @@
-{
-  "$schema": "https://docs.renovatebot.com/renovate-schema.json",
-  "extends": [
-    "config:base"
-  ],
-  "packageRules": [
-    {
-      "matchUpdateTypes": ["minor", "patch", "pin", "digest"],
-      "automerge": true
-    }
-  ],
-  "platformAutomerge": true,
-  "ignorePresets": [
-    "workarounds:javaLTSVersions"
-  ],
-  "customManagers": [
-    {
-      "customType": "regex",
-      "fileMatch": [
-        "\\.java-version$"
-      ],
-      "matchStrings": [
-        "(?<currentValue>.*)\\n"
-      ],
-      "datasourceTemplate": "java-version",
-      "depNameTemplate": "java",
-      "extractVersionTemplate": "^(?<version>\\d+)"
-    }
-  ]
-}
